+++ CHANGELOG.md
@@ -31,6 +31,18 @@ vep release_103.1 -> release_104.3
 - [NEW] gnomad mt: r3.1
 - clinvar: 20210415 -> 20211010
 
+## [10.2.5]
+
+- Allow slurm quality of service flag to be set to 'express'
+
+## [10.2.4]
+
+- Split Star-Fusion alignment and detection into two recipes
+- Use temp directory with Star-Fusion
+- Resource bump for RNA
+- Limit memory for glnexus
+- Use non-gpu version of Deepvariant by default
+
 ## [10.2.3]
 
 - Updates Chromograph to version 1.1.4